dev-requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
fastapi==0.95.0
frozenlist==1.3.3
h11==0.14.0
idna==3.4
ipykernel==6.22.0
ipython==8.12.0
isort==5.12.0
jedi==0.18.2
jupyter_client==8.1.0
jupyter_core==5.3.0
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
openai==0.27.2
packaging==23.0
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.7
Pygments==2.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pyzmq==25.0.2
requests==2.28.2
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.26.1
tomli==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
wcwidth==0.2.6
yarl==1.8.2
